i3: Update for 4.20's include directive
i3-config-gen itself is no longer needed since we can include files
now, though the workspace generator is still pretty useful.
A thrown wrench is that the include directive is not like a
C #include: i3 include will parse the included file. This results
in variables declared in one config file not making it to others:
as a result I had make a workaround where the main config file
is generated locally, outside of the repo, since the monitor variables
need to be used by the workspaces.